+++ example/lib/openapi_config.dart
@@ -2,9 +2,9 @@ import 'package:openapi_generator_annotations/openapi_generator_annotations.dart
 
 @Openapi(
     additionalProperties:
-        AdditionalProperties(pubName: 'petstore_api', pubAuthor: 'Johnny depp'),
+        AdditionalProperties(pubName: 'petstore_api', pubAuthor: 'Johnn depp'),
     inputSpecFile: 'spec/openapi-spec.yaml',
-    generatorName: 'dart',
+    generatorName: 'dart-jaguar',
     alwaysRun: true,
     outputDirectory: 'api/petstore_api')
 class OpenapiConfig extends OpenapiGeneratorConfig {